erofs-utils: mkfs: rename ctime to mtime
Currently mkfs.erofs picks up whatever the system time happened to be
when the input file structure was created. Since there's no (easy) way for
userspace to control ctime, there's no way to control the per-file ctime
that mkfs.erofs uses.

In preparation for switching to mtime, rename the "ctime" members of the
inode structure.

Link: https://lore.kernel.org/r/20220311041724.3107622-1-dvander@google.com
Signed-off-by: David Anderson <dvander@google.com>
Signed-off-by: Gao Xiang <hsiangkao@linux.alibaba.com>
Change-Id: Iff6f5f8a1e5f729f6a5949719b1891a91fd34c41